Initial Version
client_app.py -- runs on the computer controlling the car.  It will use a game PS2-style game controller if you have one, or will run a virtual set of on-screen joysticks.
car_ctl.py -- runs on the car's Pi.  Contains the functions to control the motors and servos.  If called by itself, it will run a series of tests that cycle the servos and motors
game_controller.py -- uses the pygame library to interface a PS2-style USB game controller with the client_app
PCA9685.py -- the only code I kept from the Sunfounder supplied stuff (runs on the pi).  It's messy, but works, and I didn't feel like cleaning it up
reset -- initializes the servos to the zero position and stops the motors (runs on the pi). This is useful if something crashes and leaves the car running
testcar.py -- test code (runs on the pi)
virtual_controller.py -- code for implementing a tkinter-based on-screen controller for use if you don't have a hardware game controller
tcpserver.py -- server to run on the pi.  listens for commands on the tcp port and parses/passes those commands to the car_ctl functions
clinet.cfg -- config file for the client app
config -- config file for the server (runs on the pi)